requirements.txt
# GoodData Python SDK packages
gooddata_sdk>=1.52
gooddata-pipelines>=1.52.1.dev3

pyyaml